refactor: changed file extensions from ".js to .jsx"
